// components/types.tsx
export interface Tokens {
  tokens: {
    total: number;
    tokens: Token[];
  };
}

export interface Token {
  collectionAddr: string;
  name: string;
  price?: Price;
  priceExpiresAt?: string;
  saleType?: string;
  tokenId: string;
  highestCollectionBid: string;
  description: string;
  forSale: boolean;
  imageUrl: string;
  rarityOrder: number;
  traits: Trait[];
  id: string;
  lastSale: LastSale;
  highestCollectionBidEvent: {
    bidder: {
      addr: string;
    };
  };
}

interface LastSale {
  eventName: string;
  price: Price;
}

export interface Trait {
  name: string;
  value: string;
  rarityPercent: number;
}

interface Price {
  denom: string;
  amount: string;
}